Extract _cli.py from _legacy.py [step 1 of N]
Move 4 CLI argument normalization functions (_unwrap_doublewrap,
normalize_cli_args_inplace, strip_outer_quotes, normalize_flag_values)
into a focused _cli.py submodule.

Also includes the _get_defaults_root() fallback fix for the template
gap analyzer (ensures tests find fixture defaults on a fresh checkout).

_legacy.py: -65 lines (3674 remaining).
